Split out CMSRoutes into cmswww
